% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archimedean-spirals.r
\name{arch-spirals}
\alias{arch-spirals}
\alias{sample_arch_spiral}
\title{Sample from archimedean spirals}
\usage{
sample_arch_spiral(n, min_wrap = 0, max_wrap = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{min_wrap}{The wrap of the spiral from which sampling begins}

\item{max_wrap}{The wrap of the spiral where sampling ends}
}
\description{
These functions generate uniform samples from archimedean
spirals in 2-dimensional space.
}
\details{
(Details.)
}
\examples{
#Uniformly sampled archimedean spiral in 2-space, with 2 wraps
x <- sample_arch_spiral(720, min_wrap = 0, max_wrap = 2)
plot(x, asp = 1, pch = 19, cex = .5)
}
