% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discs.r
\name{discs}
\alias{discs}
\alias{sample_disc}
\title{Sample (with noise) from disc}
\usage{
sample_disc(n, sd)
}
\arguments{
\item{n}{Number of observations.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
These functions generate uniform samples from a disc in
2-dimensional space, optionally with noise.
}
\details{
The uniform sample is generated by an area-preserving
parameterization of the disc. This parametrization was derived through the
method for sampling 2-manifolds as described by Arvo (2001).
}
\examples{
# Uniformly sampled unit disc in 2-space
x <- sample_disc(1800, sd = 0)
plot(x, asp = 1, pch = 19, cex = .5)

# Uniformly sampled unit disc in 2-space with Gaussian noise
x <- sample_disc(1800, sd = .1)
plot(x, asp = 1, pch = 19, cex = .5)
}
\references{
J Arvo (2001) Stratified Sampling of 2-Manifolds. \emph{SIGRAPH 2001 (State of the
Art in Monte Carlo Ray Tracing for Realistic Image Synthesis)}, Course Notes,
Vol. 29. \url{http://www.cs.virginia.edu/~jdl/bib/globillum/arvo01_notes.pdf}
}
