% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.r
\name{disks}
\alias{disks}
\alias{sample_disk}
\title{Sample (with noise) from disk}
\usage{
sample_disk(n, bins = 1L, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{bins}{Number of intervals per dimension to stratify by. Default set to
1, which generates a uniform sample.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
These functions generate uniform samples from a disk in
2-dimensional space, optionally with noise.
}
\details{
The sample is generated by an area-preserving parameterization of
the disk. This parametrization was derived through the method for sampling
2-manifolds as described by Arvo (2001).
}
\examples{
# Uniformly sampled unit disk in 2-space
x <- sample_disk(1800, sd = 0)
plot(x, asp = 1, pch = 19, cex = .5)

# Uniformly sampled unit disk in 2-space with Gaussian noise
x <- sample_disk(1800, sd = .1)
plot(x, asp = 1, pch = 19, cex = .5)
}
\references{
J Arvo (2001) Stratified Sampling of 2-Manifolds. \emph{SIGRAPH 2001 (State of the
Art in Monte Carlo Ray Tracing for Realistic Image Synthesis)}, Course Notes,
Vol. 29. \url{http://www.cs.virginia.edu/~jdl/bib/globillum/arvo01_notes.pdf}
}
