% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-eights.r
\name{figure-eights}
\alias{figure-eights}
\alias{sample_figure_eight}
\title{Sample from figure eight}
\usage{
sample_figure_eight(n, r)
}
\arguments{
\item{n}{Number of observations.}

\item{r}{Scalar factor of figure eight.}
}
\description{
These functions generate uniform samples from figure eights in
2-dimensional space.
}
\details{
(Details.)
}
\examples{
#Uniformly sampled figure eight in 2-space
x <- sample_figure_eight(720, 1)
plot(x, asp = 1, pch = 19, cex = .5)
}
