% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-bottle.r
\name{klein-bottle}
\alias{klein-bottle}
\alias{sample_klein}
\title{Sample (with noise) from a Klein bottle}
\usage{
sample_klein(n, r = 0.25, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
This function generates uniform samples from a Klein bottle in
  4-dimensional space, optionally with noise.
}
\details{
(Details.)
}
\examples{
inst/examples/ex-klein-bottle.r
}
