% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.r
\name{noise}
\alias{noise}
\alias{add_noise}
\title{Add noise to a sample}
\usage{
add_noise(x, sd = 0)
}
\arguments{
\item{x}{A matrix of row coordinates.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
This function adds Gaussian noise to coordinate data contained
in a matrix. It is called by samplers to introduce noise when \code{sd} is
passed a positive value.
}
