% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.r
\name{sphere}
\alias{sphere}
\alias{sample_sphere}
\title{Sample (with noise) from a sphere}
\usage{
sample_sphere(n, d = 1, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{d}{Dimension of the sphere.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
This function generates uniform samples from a sphere of radius
  1 and arbitrary dimension in 1-higher-dimensional space, optionally with
  noise.
}
\details{
This function is adapted from [TDA::sphereUnif()].
}
\examples{
inst/examples/ex-sphere.r
}
