% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirals.r
\name{spirals}
\alias{spirals}
\alias{sample_spirals}
\title{Sample (with noise) from a spirals}
\usage{
sample_spirals(n, s = 3, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{s}{Number of spirals.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
This function generates uniform samples from configurations of
  spirals in 2-dimensional space, optionally with noise.
}
\details{
This function is adapted from [KODAMA::spirals()].
}
\examples{
inst/examples/ex-spirals.r
}
