% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified-samplers.r
\name{stratified-samplers}
\alias{stratified-samplers}
\alias{stratified_square}
\title{Stratified sample of unit square}
\usage{
stratified_square(n, k)
}
\arguments{
\item{n}{Number of observations.}

\item{k}{Number of intervals per dimension for the stratification.}
}
\description{
This functions generates stratified samples of the unit square
in 2-dimensional space.
}
\details{
(Details.)
}
\examples{
#Stratified sample of a unit square with 100 cells
x <- stratified_square(110,10)
plot(x, asp = 1, pch = 19, cex = .5)
segments(x0 = seq(0,1,.1),y0 = 0, y1 = 1)
segments(y0 = seq(0,1,.1),x0 = 0, x1 = 1)
}
