% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss-roll.r
\name{swiss-roll}
\alias{swiss-roll}
\alias{sample_swiss}
\title{Sample (with noise) from a Swiss roll}
\usage{
sample_swiss(n, w = 2, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{w}{Width of the roll.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
This function generates uniform samples from a Swiss roll in
  3-dimensional space, optionally with noise.
}
\details{
This function is adapted from
  [drtoolbox](https://lvdmaaten.github.io/drtoolbox/).
}
\examples{

# Swiss roll in 3-space
x <- sample_swiss(120, sd = .1)
pairs(x, asp = 1, pch = 19, cex = .5)
}
