% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdaunif.r
\docType{package}
\name{tdaunif}
\alias{tdaunif}
\alias{tdaunif-package}
\title{`tdaunif`: Uniform manifold samplers for topological data analysis}
\description{
Generate uniform random samples from embedded manifolds,
  optionally with noise.
}
\author{
Jason Cory Brunson
}
