% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tori.r
\name{tori}
\alias{tori}
\alias{sample_torus}
\alias{sample_tori_interlocked}
\title{Sample (with noise) from tori}
\usage{
sample_torus(n, r = 0.25, sd = 0)

sample_tori_interlocked(n, r = 0.25, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{r}{Radius (relative to unit primary radius).}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
These functions generate uniform samples from configurations of
  tori of primary radius 1 in 3-dimensional space, optionally with noise.
}
\details{
(Details.)
}
\examples{
inst/examples/ex-tori.r
}
