% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trefoil-knot.R
\name{trefoil}
\alias{trefoil}
\alias{sample_trefoil}
\title{Sample from trefoil knot}
\usage{
sample_trefoil(n)
}
\arguments{
\item{n}{Number of observations.}
}
\description{
These functions generate uniform samples from trefoil knot in
3-dimensional space.
}
\details{
The trefoil knot is the simplest nontrivial knot and contains three
unique crossings in three dimensional space. This uniform sample is
generated by rejection sampling. This process allows for simulation of
random samples from the trefoil knot distribution, by using random samples
from a more convenient distribution. It applies rejection/acceptance
criterion such that the samples that are accepted are to be distributed as
if they were from the target distribution. This function is adapted from
Diaconis &al, 2013.
}
\examples{
# Uniformly sampled trefoil knot in 3-space
x <- sample_trefoil(180)
pairs(x, asp = 1, pch = 19, cex = .5, col = "#00000077")
}
