% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-discs.r
\name{unit-discs}
\alias{unit-discs}
\alias{sample_unit_disc}
\title{Sample (with noise) from unit disc}
\usage{
sample_unit_disc(n, sd)
}
\arguments{
\item{n}{Number of observations.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
These functions generate uniform samples from unit disc in
2-dimensional space, optionally with noise.
}
\details{
(Details.)
}
\examples{
# Uniformly sampled unit disc in 2-space
x <- sample_unit_disc(1800, sd = 0)
plot(x, asp = 1, pch = 19, cex = .5)

# Uniformly sampled unit disc in 2-space with Gaussian noise
x <- sample_unit_disc(1800, sd = .1)
plot(x, asp = 1, pch = 19, cex = .5)
}
\references{
J Arvo (2001) Stratified Sampling of 2-Manifolds. \emph{SIGRAPH 2001 (State of the
Art in Monte Carlo Ray Tracing for Realistic Image Synthesis)}, Course Notes,
Vol. 29. \url{http://www.cs.virginia.edu/~jdl/bib/globillum/arvo01_notes.pdf}
}
