% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archimedean-spirals.r
\name{arch-spirals}
\alias{arch-spirals}
\alias{sample_arch_spiral}
\title{Sample from archimedean spirals}
\usage{
sample_arch_spiral(n, min_wrap = 0, max_wrap = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{min_wrap}{The wrap of the spiral from which sampling begins}

\item{max_wrap}{The wrap of the spiral where sampling ends}
}
\description{
These functions generate uniform samples from archimedean
spirals in 2-dimensional space.
}
\details{
The archimedean spiral starts at the origin and wraps around itself
such that distances between all the spiral branches are equivalent. The
specific parametrization was taken from (insert). The uniform sample is
generated through a rejection sampling process as described by Diaconis
(2013).
}
\examples{
#Uniformly sampled archimedean spiral in 2-space, with 1 wrap
x <- sample_arch_spiral(360, min_wrap = 0, max_wrap = 1)
plot(x, asp = 1, pch = 19, cex = .5)

#Uniformly sampled archimedean spiral in 2-space, with 5 wraps
x <- sample_arch_spiral(360, min_wrap = 0, max_wrap = 5)
plot(x, asp = 1, pch = 19, cex = .5)

#Uniformly sampled archimedean spiral in 2-space, with 5 wraps and starting from
#2 wraps
x <- sample_arch_spiral(360, min_wrap = 2, max_wrap = 5)
plot(x, asp = 1, pch = 19, cex = .5)
}
\references{
P Diaconis, S Holmes, and M Shahshahani (2013) Sampling from a Manifold.
\emph{Advances in Modern Statistical Theory and Applications: A Festschrift in
honor of Morris L. Eaton}, 102--125. \doi{10.1214/12-IMSCOLL1006}
}
