% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.r
\name{circles}
\alias{circles}
\alias{sample_circle}
\alias{sample_circles_interlocked}
\alias{sample_circle_sinusoidal}
\alias{sample_eight}
\title{Sample (with noise) from circles}
\usage{
sample_circle(n, sd = 0)

sample_circles_interlocked(n, sd = 0)

sample_circle_sinusoidal(n, sd = 0)

sample_eight(n, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
These functions generate uniform samples from configurations of
  circles of radius 1 in 2- or 3-dimensional space, optionally with noise.
}
\details{
(Details.)
}
\examples{
inst/examples/ex-circles.r
}
