% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-bottles.r
\name{klein-bottles}
\alias{klein-bottles}
\alias{sample_klein_tube}
\alias{sample_klein}
\alias{sample_klein_flat}
\title{Sample (with noise) from Klein bottles}
\usage{
sample_klein_tube(n, ar = 2, sd = 0)

sample_klein(n, ar = 2, sd = 0)

sample_klein_flat(n, ar = 1, bump = 0.1, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{ar}{Aspect ratio for Möbius tube Klein bottle (ratio of major and minor
radii) or flat torus-based Klein bottle (ratio of scale factors).}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}

\item{bump}{Bump constant for the flat torus-based Klein bottle.}
}
\description{
These functions generate uniform samples from Klein bottles in
  4-dimensional space, optionally with noise.
}
\details{
(Details.)
}
\examples{
inst/examples/ex-klein-bottle.r
}
