% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real-projective-plane.r
\name{real-projective-planes}
\alias{real-projective-planes}
\alias{sample_projective_plane}
\title{Sample (with noise) from real projective planes}
\usage{
sample_projective_plane(n, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
These functions generate uniform samples from real projective
planes in 4-dimensional space, optionally with noise.
}
\details{
(Details.)
}
