% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.r
\name{sphere}
\alias{sphere}
\alias{sample_2sphere}
\alias{sample_sphere}
\title{Sample (with noise) from a sphere}
\usage{
sample_2sphere(n, sd = 0)

sample_sphere(n, d = 1, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}

\item{d}{Dimension of the sphere.}
}
\description{
These functions generate uniform samples from a sphere of radius
  1 and dimension 2 in 3-space, or in arbitrary dimension in
  1-higher-dimensional space, optionally with noise.
}
\details{
The function `sample_sphere` is adapted from [TDA::sphereUnif()].
}
\examples{

# 1-sphere in 2-space
x <- sample_sphere(120, d = 1, sd = .1)
pairs(x, asp = 1, pch = 19, cex = .5)

# 2-sphere in 3-space
x <- sample_sphere(120, d = 2, sd = .1)
pairs(x, asp = 1, pch = 19, cex = .5)

# 3-sphere in 4-space
x <- sample_sphere(120, d = 3, sd = .1)
pairs(x, asp = 1, pch = 19, cex = .5)

# 4-sphere in 5-space
x <- sample_sphere(120, d = 4, sd = .1)
pairs(x, asp = 1, pch = 19, cex = .5)
}
