% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirals.r
\name{spirals}
\alias{spirals}
\alias{sample_spirals}
\title{Sample (with noise) from a spirals}
\usage{
sample_spirals(n, spirals = 3, sd = 0)
}
\arguments{
\item{n}{Number of observations.}

\item{spirals}{Number of spirals.}

\item{sd}{Standard deviation of (independent multivariate) Gaussian noise.}
}
\description{
This function generates uniform samples from configurations of
spirals in 2-dimensional space, optionally with noise.
}
\details{
This function is adapted from \code{\link[KODAMA:spirals]{KODAMA::spirals()}}.

\strong{Note:} This is a non-uniform sampler. Use \code{\link[=sample_arch_spiral]{sample_arch_spiral()}}
instead, which will be modified soon to incorporate multiple arms.
}
\examples{

# 1 spiral in 2-space
x <- sample_spirals(120, s = 1, sd = .2)
plot(x, asp = 1, pch = 19, cex = .5)

# 3 spirals in 2-space
x <- sample_spirals(120, s = 3, sd = .2)
plot(x, asp = 1, pch = 19, cex = .5)
}
